% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zscore_function.R
\name{COMPUTEZ}
\alias{COMPUTEZ}
\title{COMPUTEZ function}
\usage{
COMPUTEZ(input_data, tract, patient_ID, HC = 0, ID_column, Group_column)
}
\arguments{
\item{input_data}{A dataframe containing tract data}

\item{tract}{Column name of raw tract data}

\item{patient_ID}{The column index corresponding to patient ID}

\item{HC}{Healthy control identifier. Defaults to 0}

\item{ID_column}{The column index corresponding to participant IDs. Defaults to 1}

\item{Group_column}{The column index corresponding to the Group identifier. Defaults to 12}
}
\description{
This function computes the mean and SD on tract data across all healthy control
participants included in the data. Z-scores are then computed for a specified
patient, compared to the control cohort.
}
\examples{
COMPUTEZ()
}
