% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zscore_function.R
\name{RANKINT}
\alias{RANKINT}
\title{RANKINT function}
\usage{
RANKINT(input_data, tract, int_tract)
}
\arguments{
\item{input_data}{A dataframe containing tract data.}

\item{tract}{Column name of tract data to be normalised}

\item{int_tract}{Column name of inverse-normal transformed tract data}
}
\description{
This function performs rank-based inverse normal transformation (INT) on tract
data using the RNOmni 'RankNorm' function.
}
\examples{
RANKINT()
}
