% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zscore_function.R
\name{RUNTRACTZ}
\alias{RUNTRACTZ}
\title{RUNTRACTZ function}
\usage{
RUNTRACTZ(
  input_data,
  tractrange,
  patient_ID,
  combat = TRUE,
  ID_column = 1,
  scanner_column = 10,
  age_column = 8
)
}
\arguments{
\item{input_data}{A dataframe containing tract data}

\item{tractrange}{The column(s) corresponding to tract data}

\item{patient_ID}{The column index corresponding to patient ID}

\item{combat}{Perform neuroCombat site/scanner harmonisation? Defaults to TRUE}

\item{ID_column}{The column index corresponding to participant IDs. Defaults to 1}

\item{age_column}{The column index corresponding to age. Defaults to 8}
}
\description{
This function performs the computation of tract-specific Z-scores for a given patient,
using the RANKINT and COMPUTEZ functions. A dataframe should be provided as input,
with a column corresponding to patient ID, and column(s) corresponding to tract-specific
measures included as a minimum. If data harmonisation is required, columns corresponding
to site (numeric identifier) and age should also be included
}
\examples{
RUNTRACTZ()
}
