% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider_function.R
\name{SPIDERZ}
\alias{SPIDERZ}
\title{SPIDERZ function}
\usage{
SPIDERZ(
  input_Zscores,
  bilateral = TRUE,
  sigline = TRUE,
  customtracts = FALSE,
  leftcol = "#3498DB",
  rightcol = "#E74C3C"
)
}
\arguments{
\item{input_Zscores}{A dataframe containing Z-scores for tract data for a patient.}

\item{bilateral}{I have not done this yet. Defaults to TRUE.}

\item{sigline}{Include a significance threshold line (Z=-1.96) in plot? Defaults to TRUE}

\item{customtracts}{I have not done this yet. Defaults to FALSE}

\item{leftcol}{Colour for left-sided tracts. Defaults to "#3498DB"}

\item{rightcol}{Colour for left-sided tracts. Defaults to "#E74C3C"}
}
\description{
This function plots tract-based data as a radar (or spider) chart. Input data
should be tract-specific Z-scores for a given patient, when compared to a
control cohort. Z-scores can be computed using the RUNTRACTZ function if input
data contains tract data for both healthy control participants and the patient.
}
\examples{
SPIDERZ()
}
