% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tractdata}
\alias{tractdata}
\title{Tract FDC data from fixel-based analysis of DWI data}
\format{
A data frame with 100 rows and 36 variables:
\describe{
\item{Group}{Group of participant, either patient (1) or control (0)}
\item{Scanner_ID}{identifier of scanner or site, for use in ComBat}
...
}
}
\usage{
data(tractdata)
}
\description{
Data from a cohort of healthy control participants and
patients with epilepsy. The data included are mean fibre density
and cross-section (FDC) values from selected tracts of interest,
and were computed using the fixel-based analysis pipeline.
}
\keyword{datasets}
